Create UI framework and automate loading (most) fonts
