+++ OstelcoStyles/OstelcoStyles.h
@@ -1,0 +1,19 @@
+//
+//  OstelcoStyles.h
+//  OstelcoStyles
+//
+//  Created by Ellen Shapiro on 4/23/19.
+//  Copyright © 2019 mac. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for OstelcoStyles.
+FOUNDATION_EXPORT double OstelcoStylesVersionNumber;
+
+//! Project version string for OstelcoStyles.
+FOUNDATION_EXPORT const unsigned char OstelcoStylesVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <OstelcoStyles/PublicHeader.h>
+
+
