+++ ostelco-ios-client.xcodeproj/xcshareddata/xcschemes/OstelcoStyles.xcscheme
@@ -14,37 +14,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "9B851ABA225E29BE0076ABF3"
-               BuildableName = "ostelco_core.framework"
-               BlueprintName = "ostelco-core"
-               ReferencedContainer = "container:ostelco-ios-client.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
                BlueprintIdentifier = "9BB6A792226F6513000D5E13"
                BuildableName = "OstelcoStyles.framework"
                BlueprintName = "OstelcoStyles"
-               ReferencedContainer = "container:ostelco-ios-client.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "04488165217924B700F59758"
-               BuildableName = "Oya.app"
-               BlueprintName = "ostelco-ios-client"
                ReferencedContainer = "container:ostelco-ios-client.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -57,15 +29,6 @@
       shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
       </Testables>
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "04488165217924B700F59758"
-            BuildableName = "Oya.app"
-            BlueprintName = "ostelco-ios-client"
-            ReferencedContainer = "container:ostelco-ios-client.xcodeproj">
-         </BuildableReference>
-      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </TestAction>
@@ -79,16 +42,15 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "04488165217924B700F59758"
-            BuildableName = "Oya.app"
-            BlueprintName = "ostelco-ios-client"
+            BlueprintIdentifier = "9BB6A792226F6513000D5E13"
+            BuildableName = "OstelcoStyles.framework"
+            BlueprintName = "OstelcoStyles"
             ReferencedContainer = "container:ostelco-ios-client.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
@@ -98,16 +60,15 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "04488165217924B700F59758"
-            BuildableName = "Oya.app"
-            BlueprintName = "ostelco-ios-client"
+            BlueprintIdentifier = "9BB6A792226F6513000D5E13"
+            BuildableName = "OstelcoStyles.framework"
+            BlueprintName = "OstelcoStyles"
             ReferencedContainer = "container:ostelco-ios-client.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
